requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
pyperclip>=1.8.2
requests>=2.31.0
tkinter>=8.6 